+++ package.json
@@ -49,6 +49,6 @@
     "gatsby-remark-prismjs": "^1.2.8",
     "gatsby-source-filesystem": "^1.5.5",
     "gatsby-transformer-remark": "^1.7.17",
-    "typography-theme-github": "^0.15.10"
+    "typography-theme-moraga": "^0.15.10"
   }
 }
