+++ src/utils/typography.js
@@ -1,5 +1,5 @@
 import Typography from 'typography';
-import theme from 'typography-theme-github';
+import theme from 'typography-theme-moraga';
 
 theme.baseFontSize = `20px`;
 const typography = new Typography(theme);
